# FintechCopilot.Application/DTOs/AuthResponse.cs
namespace FintechCopilot.Application.DTOs;

public class AuthResponse
{
    public string Token { get; set; } = string.Empty;
    public string CompanyName { get; set; } = string.Empty;
    public string Email { get; set; } = string.Empty;
}