# FintechCopilot.Application/DTOs/LoginRequest.cs
using System.ComponentModel.DataAnnotations;

namespace FintechCopilot.Application.DTOs;

public class LoginRequest
{
    [Required(ErrorMessage = "El correo electrónico es obligatorio")]
    [EmailAddress(ErrorMessage = "Formato de correo inválido")]
    public string Email { get; set; } = string.Empty;

    [Required(ErrorMessage = "La contraseña es obligatoria")]
    public string Password { get; set; } = string.Empty;
}