# openapi.yaml
openapi: 3.0.3
info:
  title: FintechCopilot API
  description: API para el Copiloto de Análisis Financiero y Facturación Inteligente.
  version: 1.0.0
paths:
  /api/auth/register:
    post:
      summary: Registrar una nueva empresa/usuario en el sistema.
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registro exitoso. Devuelve el token JWT y datos del usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/auth/login:
    post:
      summary: Iniciar sesión en el sistema.
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login exitoso. Devuelve el token JWT y datos del usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Credenciales inválidas.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Datos inválidos o el correo ya existe.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/invoices/upload:
    post:
      summary: Subir una factura o recibo (PDF/Imagen) para extracción por IA.
      tags:
        - Invoices
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Archivo procesado exitosamente por la IA.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          description: Archivo inválido (formato o tamaño no permitido).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/invoices/confirm:
    post:
      summary: Confirmar y guardar los datos extraídos de la factura.
      tags:
        - Invoices
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmInvoiceRequest'
      responses:
        '200':
          description: Gasto guardado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmInvoiceResponse'
  /api/invoices:
    get:
      summary: Obtener el historial de gastos del usuario.
      tags:
        - Invoices
      responses:
        '200':
          description: Lista de gastos registrados.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvoiceResponse'
  /api/invoices/{id}:
    delete:
      summary: Eliminar una factura por su ID.
      tags:
        - Invoices
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Gasto eliminado correctamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Gasto eliminado correctamente"
        '404':
          description: Factura no encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    RegisterRequest:
      type: object
      required:
        - companyName
        - email
        - password
      properties:
        companyName:
          type: string
          example: "Mi Empresa S.A.S"
        email:
          type: string
          format: email
          example: "contacto@miempresa.com"
        password:
          type: string
          format: password
          minLength: 8
          example: "Segura123!"
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "contacto@miempresa.com"
        password:
          type: string
          format: password
          example: "Segura123!"
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        companyName:
          type: string
          example: "Mi Empresa S.A.S"
        email:
          type: string
          example: "contacto@miempresa.com"
    UploadResponse:
      type: object
      properties:
        invoiceId:
          type: string
          format: uuid
          example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
        fileName:
          type: string
          example: "factura_001.pdf"
        extractedData:
          $ref: '#/components/schemas/ExtractedInvoiceData'
    ExtractedInvoiceData:
      type: object
      properties:
        provider:
          type: string
          example: "Proveedor S.A.S"
        date:
          type: string
          format: date
          example: "2026-06-01"
        total:
          type: number
          format: float
          example: 1250000.00
        tax:
          type: number
          format: float
          example: 237500.00
        currency:
          type: string
          example: "COP"
        category:
          type: string
          example: "Servicios"
    ConfirmInvoiceRequest:
      type: object
      required:
        - invoiceId
        - provider
        - date
        - total
        - tax
        - category
      properties:
        invoiceId:
          type: string
          format: uuid
        provider:
          type: string
          example: "AWS"
        date:
          type: string
          format: date
          example: "2026-06-01"
        total:
          type: number
          example: 120.50
        tax:
          type: number
          example: 19.28
        currency:
          type: string
          example: "USD"
        category:
          type: string
          example: "Tecnología"
    ConfirmInvoiceResponse:
      type: object
      properties:
        message:
          type: string
          example: "Gasto guardado correctamente"
    InvoiceResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        provider:
          type: string
        date:
          type: string
          format: date
        total:
          type: number
        tax:
          type: number
        currency:
          type: string
        category:
          type: string
        createdAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Este correo ya está registrado, intenta iniciar sesión"